backend/requirements.txt
uvicorn==0.30.6
fastapi==0.112.0
PyJWT==2.9.0
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.32